requirements.txt
flask
python-dotenv
azure-identity
azure-storage-blob
azure-storage-queue
PyPDF2
pdf2image
Pillow
uuid
azure-ai-documentintelligence
flask-cors
pyautogen
pytest
opentelemetry-instrumentation-flask
langchain
langchain-openai
tenacity
flask-socketio
gunicorn
gevent
graphrag==0.1.1
pandas
adlfs
fsspec
azure-data-tables